Provide support for JSON Web Tokens instead of session-cookie auth

+++ server/api.js
@@ -12,91 +12,105 @@ const ogScraper = require('open-graph-scraper');
 const passport = require('passport');
 // const ineed = require('ineed');
 
+const { apiError } = require('./api-error');
 const DB = require('./db');
 const auth = require('./auth');
 const transform = require('../common/transforms');
 const { resources, routes, queryLimit } = require('../common/enums.api.js');
 
-// @todo: log errors server-side
-function errorHandler (err, errCode, msg, res) {
-    const errName = 'API Error';
-    const errResponse = { msg: `${errName}: ${msg}`, status: errCode, error: err.message };
-    res.status(errCode);
-    res.json(errResponse);
-}
-
 /**
- * Validate requests
- * @param req
- * @param res
- * @param next
+ * ValidateUser middleware
+ * Checks that the user session is valid
  */
-function validator (req, res, next) {
-    // @todo: SECURITY validate access
-    // @todo: SECURITY handle/cleanse request data
-    const { resource } = req.params;
+function validateUser (req, res, next) {
+    console.info('User', req.user, res.user, res);
 
-    if (resource && !_.includes(resources, resource)) {
-        return errorHandler({}, 400, 'Invalid resource', res);
+    if (!req.user || req.user.length > 1) {
+        return apiError({ message: 'Not authenticated' }, 401, 'Not authenticated', res);
     }
+
+    req.userID = req.user[0]._id;
+
     next();
 }
 
 /**
- * Get User
- * @param req
- * @param res
- * @param id
+ * ValidateResource middleware
+ * Checks that the requested resource is valid
  */
-function getUserObject (req, res, id) {
-    (id) ? res.json({ username: id }) : errorHandler(req, 500, 'User not found', res);
+function validateResource (req, res, next) {
+    // @todo: SECURITY handle/cleanse request data
+    const { resource } = req.params;
+
+    if (resource && !_.includes(resources, resource)) {
+        apiError({}, 400, 'Invalid resource', res);
+    }
+    next();
 }
 
 module.exports = (app) => {
-    app.post(routes.auth.login, csrf, passport.authenticate('local'), (req, res) => {
-        // If this function gets called, authentication was successful.
-        res.json({ success: true });
+
+    /**
+     * API: POST Auth/Login
+     * User login
+     */
+    app.post(routes.auth.login,
+        csrf,
+        passport.authenticate('local'),
+        auth.loginUser,
+        (req, res) => {
+            // Authentication successful.
+            res.json({ success: true });
     });
 
     /**
      *  API: POST User
      *  Create new user accounts
      */
-    app.post(routes.user, csrf, (req, res) => {
-        const { email, password } = req.body;
-        if (email && password) {
-            auth.createUser({ email, password })
-                .then((result) => res.json(result.ops))
-                .catch((err) => {
-                    console.info('createUser error', err);
-                    errorHandler(err, 409, 'Could not create user', res);
-                });
-        } else {
-            errorHandler({}, 400, 'Nothing to do: "email" and/or "password" parameters missing.', res);
+    app.post(routes.user,
+        csrf,
+        (req, res) => {
+            const { email, password } = req.body;
+            if (email && password) {
+                auth.createUser({ email, password })
+                    .then((result) => res.json(result.ops))
+                    .catch((err) => {
+                        console.info('createUser error', err);
+                        apiError(err, 409, 'Could not create user', res);
+                    });
+            } else {
+                apiError({}, 400, 'Nothing to do: "email" and/or "password" parameters missing.', res);
+            }
         }
-    });
+    );
 
     /**
      *  API: GET CSRF Token
      */
-    app.get(routes.auth.token, csrf, validator, (req, res) => res.json({ token: req.csrfToken() }) );
+    app.get(routes.auth.token,
+        csrf,
+        (req, res) => res.json({ token: req.csrfToken() })
+    );
 
     /**
      * API: GET Product URL
      * Fetch URL data: OG Tags/Scraped Data
      */
-    app.get(routes.product, validator, (req, res) => {
-        const url = decodeURIComponent(req.params.url);
-        const resultDefaults = { opengraph: false, scraped: false };
+    app.get(routes.product,
+        auth.verifyUser,
+        (req, res) => {
+            const url = decodeURIComponent(req.params.url);
+            const resultDefaults = { opengraph: false, scraped: false };
 
-        // Scrape from OpenGraph tags
-        ogScraper({ url, timeout: 5000 }, (err, result) => {
-            if (err) return errorHandler(err, 500, `Could not collect product resources from: ${url}. Reason: ${result.err}`, res);
-            res.json(_.extend(result, resultDefaults, { opengraph: true }));
-        });
+            // Scrape from OpenGraph tags
+            ogScraper({ url, timeout: 5000 }, (err, result) => {
+                if (err) return apiError(err, 500, `Could not collect product resources from: ${url}. Reason: ${result.err}`, res);
+                res.json(_.extend(result, resultDefaults, { opengraph: true }));
+            });
 
-        // @todo: Scrape for data if OG returns no useful results
-    });
+            // @todo: Scrape for data if OG returns no useful results
+        }
+    );
 
     /**
      * API: POST Resource/Collection
@@ -104,30 +118,33 @@ module.exports = (app) => {
      *
      * @greedy
      */
-    app.post(routes.collection, csrf, validator, (req, res) => {
-        const user = 'nijk'; // @todo: User Authentication
+    app.post(routes.collection,
+        csrf,
+        validateResource,
+        auth.verifyUser,
+        (req, res) => {
+            const { resource, collection } = req.params;
+            const { item } = req.body;
 
-        const { resource, collection } = req.params;
-        const { item } = req.body;
+            if (collection && item) {
+                DB.createDocument({ user: req.userID, resource, collection, doc: item })
+                    .then((result) => res.json(result.ops))
+                    .catch((err) => apiError(err, 500, 'Could not create document', res));
 
-        if (collection && item) {
-            DB.createDocument({ user, resource, collection, doc: item })
-                .then((result) => res.json(result.ops))
-                .catch((err) => errorHandler(err, 500, 'Could not create document', res));
-
-        } else if (item) {
-            DB.createCollection({ user, resource, collection: item.name })
-                .then((name) => {
-                    const location = transform.route(routes.collection, { resource: resource, collection: name });
-                    res.status(201);
-                    res.set('Location', location);
-                    res.send();
-                })
-                .catch((err) => errorHandler(err, 500, 'Could not create collection', res));
-        } else {
-            errorHandler({}, 500, 'Nothing to do: "item" parameter missing.', res);
+            } else if (item) {
+                DB.createCollection({ user: req.userID, resource, collection: item.name })
+                    .then((name) => {
+                        const location = transform.route(routes.collection, { resource: resource, collection: name });
+                        res.status(201);
+                        res.set('Location', location);
+                        res.send();
+                    })
+                    .catch((err) => apiError(err, 500, 'Could not create collection', res));
+            } else {
+                apiError({}, 500, 'Nothing to do: "item" parameter missing.', res);
+            }
         }
-    });
+    );
 
     /**
      * API: PUT Document
@@ -135,21 +152,24 @@ module.exports = (app) => {
      *
      * @greedy
      */
-    app.put(routes.collection, csrf, validator, (req, res) => {
-        const user = 'nijk'; // @todo: User Authentication
+    app.put(routes.collection,
+        csrf,
+        validateResource,
+        auth.verifyUser,
+        (req, res) => {
+            const { resource, collection } = req.params;
+            const { item } = req.body;
 
-        const { resource, collection } = req.params;
-        const { item } = req.body;
+            if (collection && item) {
+                DB.updateDocument({ user: req.userID, resource, collection, doc: item })
+                    .then((result) => res.json({ result, item }))
+                    .catch((err) => apiError(err, 500, 'Could not update document', res));
 
-        if (collection && item) {
-            DB.updateDocument({ user, resource, collection, doc: item })
-                .then((result) => res.json({ result, item }))
-                .catch((err) => errorHandler(err, 500, 'Could not update document', res));
-
-        } else {
-            errorHandler({}, 409, 'Nothing to do: "collection" and/or "item" parameters missing.', res);
+            } else {
+                apiError({}, 409, 'Nothing to do: "collection" and/or "item" parameters missing.', res);
+            }
         }
-    });
+    );
 
     /**
      * API: DELETE Document
@@ -157,20 +177,22 @@ module.exports = (app) => {
      *
      * @greedy
      */
-    app.delete(routes.collection, /*csrf, */validator, (req, res) => {
-        const user = 'nijk'; // @todo: User Authentication
+    app.delete(routes.collection,
+        validateResource,
+        auth.verifyUser,
+        (req, res) => {
+            const { resource, collection, id } = req.params;
 
-        const { resource, collection, id } = req.params;
+            if (collection && id) {
+                DB.removeDocument({ user: req.userID, resource, collection, id })
+                    .then((result) => res.json({ result }))
+                    .catch((err) => apiError(err, 500, 'Could not remove document', res));
 
-        if (collection && id) {
-            DB.removeDocument({ user, resource, collection, id })
-                .then((result) => res.json({ result }))
-                .catch((err) => errorHandler(err, 500, 'Could not remove document', res));
-
-        } else {
-            errorHandler({}, 409, 'Nothing to do: "collection" and/or "id" parameters missing.', res);
+            } else {
+                apiError({}, 409, 'Nothing to do: "collection" and/or "id" parameters missing.', res);
+            }
         }
-    });
+    );
 
     /**
      * API: GET Resource/Collection
@@ -178,27 +200,25 @@ module.exports = (app) => {
      *
      * @greedy
      */
-    app.get(routes.collection, validator, (req, res) => {
-        const user = 'nijk'; // @todo: User Authentication
-        const { resource, collection/*, type, id*/ } = req.params;
-        let { page, limit } = req.query;
+    app.get(routes.collection,
+        validateResource,
+        auth.verifyUser,
+        (req, res) => {
+            const { resource, collection } = req.params;
+            let { page, limit } = req.query;
+            page = parseInt(page || 1);
+            limit = parseInt(limit || queryLimit);
 
-        page = parseInt(page || 1);
-        limit = parseInt(limit || queryLimit);
+            if (collection) {
+                DB.retrieveDocuments({ user: req.userID, resource, collection, page, limit })
+                    .then(res.json)
+                    .catch((err) => apiError(err, 500, 'Could not retrieve documents', res));
+            } else {
+                DB.retrieveCollections({ user: req.userID, resource, page, limit })
+                    .then(res.json)
+                    .catch((err) => apiError(err, 500, 'Could not retrieve collections', res));
 
-        if (collection) {
-            DB.retrieveDocuments({ user, resource, collection, page, limit })
-                .then(res.json)
-                .catch((err) => errorHandler(err, 500, 'Could not retrieve documents', res));
-        } else {
-            DB.retrieveCollections({ user, resource, page, limit })
-                .then(res.json)
-                .catch((err) => errorHandler(err, 500, 'Could not retrieve collections', res));
-
+            }
         }
-    });
-
-    /*app.get('/api/user/:id?', auth.authenticate('local'), function(req, res) {
-        res.json(getUserObject(req, res, req.params.id));
-    });*/
+    );
 };
